+++ IoTSonnenUploader/.gitignore
@@ -4,3 +4,4 @@
 /.project
 /.factorypath
 /.classpath
+/OCISetup/
