--- IoTSonnenUploader/config-sample/mqtt.yml
@@ -1,40 +1,0 @@
-mqtt:
-  monitorreference:
-    enabled: false
-  monitoruploads:
-    enabled: false
-  statusupload:
-    enabled: true
-    frequency: "10s"
-    initialdelay: "10s"
-  configurationupload:
-    enabled: true
-    frequency: "10s"
-    initialdelay: "5s"
-  device:
-    id: tims
-  broker:
-    host: localhost
-    port: 1883
-  client:
-    # Unique identifier for this MQTT client
-    client-id: ${random.uuid}
-
-    # MQTT broker connection
-    server-uri: tcp://${mqtt.broker.host}:${mqtt.broker.port}
-
-    # Authentication
-    user-name: ${mqtt.device.id}
-    password: ExamplePassword
-
-    # Optional connection settings
-    clean-session: true
-    automatic-reconnect: true
-    connection-timeout: "PT30s"
-    keep-alive-interval: 60
-
-    will-message:
-      topic: home/sonnen/will/${mqtt.device.id}
-      payload: Device ${mqtt.device.id} is unavailable
-      qos: 0
-      retained: true
