+++ IoTSonnenUploader/src/main/java/com/oracle/demo/timg/iot/iotsonnenuploader/mqtt/MqttCommandResponsePublisher.java
@@ -38,6 +38,8 @@ package com.oracle.demo.timg.iot.iotsonnenuploader.mqtt;
 
 import java.util.concurrent.CompletableFuture;
 
+import com.oracle.demo.timg.iot.iotsonnenuploader.devicesettings.DeviceSettings;
+
 import io.micronaut.context.annotation.Requires;
 import io.micronaut.mqtt.annotation.Topic;
 import io.micronaut.mqtt.annotation.v5.MqttPublisher;
@@ -45,9 +47,9 @@ import io.micronaut.scheduling.TaskExecutors;
 import io.micronaut.scheduling.annotation.ExecuteOn;
 
 @MqttPublisher
-@Requires(property = MqttDeviceSettings.PREFIX + ".id")
+@Requires(property = DeviceSettings.PREFIX + ".id")
 public interface MqttCommandResponsePublisher {
-	@Topic("house/sonnen/commandresponse/${" + MqttDeviceSettings.PREFIX + ".id}")
+	@Topic("house/sonnen/commandresponse/${" + DeviceSettings.PREFIX + ".id}")
 	@ExecuteOn(TaskExecutors.IO)
 	public CompletableFuture<Void> publishCommandResponse(byte[] responsebytes);
 }
